classification/requirements.txt
torch==1.7.0
torchvision==0.8.1
timm==0.3.2
mmcv-full==1.3.8
